feat: building the implication graph
