+++ package.json
@@ -61,6 +61,7 @@
 		"vitest": "^3.0.0"
 	},
 	"dependencies": {
+		"d3": "^7.9.0",
 		"katex": "^0.16.21",
 		"serve": "^14.2.4",
 		"svelte-tiny-virtual-list": "^2.1.2"
